+cyrillic

+++ sql/db-create.sql
@@ -106,8 +106,8 @@ CREATE TABLE IF NOT EXISTS `taxi`.`receipts` (
   `id` INT NOT NULL AUTO_INCREMENT,
   `user_id` INT NOT NULL,
   `car_id` INT NOT NULL,
-  `price` DOUBLE NOT NULL,
-  `length` DOUBLE NOT NULL,
+  `price` INT NOT NULL,
+  `length` INT NOT NULL,
   `destination` VARCHAR(45) NOT NULL,
   `departure` VARCHAR(45) NOT NULL,
   `date` VARCHAR(20) NOT NULL,
@@ -137,7 +137,7 @@ CREATE TABLE IF NOT EXISTS `taxi`.`tariffs` (
   `id` INT NOT NULL,
   `types_id` INT NOT NULL,
   `price_for_km` INT NOT NULL,
-  `length` DOUBLE NOT NULL,
+  `length` INT NOT NULL,
   PRIMARY KEY (`id`))
 ENGINE = InnoDB;
